requirements.txt
# TranSIT Tag Localisation - Python Dependencies
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0

# Computer vision (QR-code module)
opencv-python>=4.5.0

# Excel file handling
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Serial communication (RFID data extraction)
pyserial>=3.5
